pyproject.toml
[project]
name = "yamlcli"
version = "0.2.0"
description = "Convert YAML files to JSON and vice versa"
requires-python = ">=3.10,<4.0"
readme = "README.md"
authors = [
    { name = "RyoNak", email = "nakagamiryo0901@gmail.com" }
]
dependencies = [
    "pyyaml",
    "typer>=0.20.0",
]

[project.scripts]
yamlcli = "yamlcli:main"

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=term-missing"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=4.0"
]
